upgrade to version 2.1
