+++ .gitlab-ci.yml
@@ -4,9 +4,9 @@ stages:
 .build-deploy:
   script:
     - 'mvn clean package'
-    - 'tar zxvf ./target/sql-runner-2.0-bin.tar.gz -C ./target'
-    - 'rm -rf /data/ws/sql-runner-2.0'
-    - 'mv ./target/sql-runner-2.0 /data/ws/sql-runner-2.0'
+    - 'tar zxvf ./target/sql-runner-2.1-bin.tar.gz -C ./target'
+    - 'rm -rf /data/ws/sql-runner-2.1'
+    - 'mv ./target/sql-runner-2.1 /data/ws/sql-runner-2.1'
   stage: deploy
   when: manual
