+++ pom.xml
@@ -6,7 +6,7 @@
 
   <groupId>org.apache.sql.runner</groupId>
   <artifactId>sql-runner</artifactId>
-  <version>2.0</version>
+  <version>2.1</version>
 
   <properties>
     <maven.compiler.encoding>UTF-8</maven.compiler.encoding>
