--- src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,2 +1,0 @@
-org.apache.spark.sql.execution.datasources.JdbcUpsertRelationProvider
-org.apache.spark.sql.execution.datasources.KafkaSinkRelationProvider
